+++ BitFaster.Caching/Lru/TelemetryPolicy.cs
@@ -1,11 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Data.Common;
-using System.Linq;
-using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
-using BitFaster.Caching.Pad;
+using BitFaster.Caching.Concurrent;
 
 namespace BitFaster.Caching.Lru
 {
@@ -13,8 +8,8 @@ namespace BitFaster.Caching.Lru
     {
         private LongAdder hitCount;
         private LongAdder missCount;
-        private long evictedCount;
-        private long updatedCount;
+        private LongAdder evictedCount;
+        private LongAdder updatedCount;
         private object eventSource;
 
         public event EventHandler<ItemRemovedEventArgs<K, V>> ItemRemoved;
@@ -27,9 +22,9 @@ namespace BitFaster.Caching.Lru
 
         public long Misses => this.missCount.Sum();
 
-        public long Evicted => this.evictedCount;
+        public long Evicted => this.evictedCount.Sum();
 
-        public long Updated => this.updatedCount;
+        public long Updated => this.updatedCount.Sum();
 
         public void IncrementMiss()
         {
@@ -45,7 +40,7 @@ namespace BitFaster.Caching.Lru
         {
             if (reason == ItemRemovedReason.Evicted)
             {
-                Interlocked.Increment(ref this.evictedCount);
+                this.evictedCount.Increment();
             }
 
             // passing 'this' as source boxes the struct, and is anyway the wrong object
@@ -54,13 +49,15 @@ namespace BitFaster.Caching.Lru
 
         public void OnItemUpdated(K key, V value)
         {
-            Interlocked.Increment(ref this.updatedCount);
+            this.updatedCount.Increment();
         }
 
         public void SetEventSource(object source)
         {
             this.hitCount = new LongAdder();
             this.missCount = new LongAdder();
+            this.evictedCount = new LongAdder();
+            this.updatedCount = new LongAdder();
             this.eventSource = source;
         }
     }
